Fix filter controls having margin when no controls given, add DisplayStyleControl to social, misc cleanups
