+++ osu.Game/Overlays/Social/FilterControl.cs
@@ -1,6 +1,7 @@
 // Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
 // Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu/master/LICENCE
 
+using System.ComponentModel;
 using OpenTK.Graphics;
 using osu.Game.Graphics;
 using osu.Game.Overlays.Browse;
@@ -11,5 +12,16 @@ namespace osu.Game.Overlays.Social
     {
         protected override Color4 BackgroundColour => OsuColour.FromHex(@"47253a");
         protected override SocialSortCriteria DefaultTab => SocialSortCriteria.Name;
+    }
+
+    public enum SocialSortCriteria
+    {
+        Name,
+        Rank,
+        Location,
+        [Description("Time Zone")]
+        TimeZone,
+        [Description("World Map")]
+        WorldMap,
     }
 }
